64-bit compatibility

+++ src/cbdata.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: cbdata.cc,v 1.68 2006/01/03 17:22:30 wessels Exp $
+ * $Id: cbdata.cc,v 1.69 2006/01/24 15:42:27 hno Exp $
  *
  * DEBUG: section 45    Callback Data Registry
  * ORIGINAL AUTHOR: Duane Wessels
@@ -517,7 +517,7 @@ cbdataDump(StoreEntry * sentry)
 
         if (pool) {
             int obj_size = pool->objectSize() - cbdata::Offset;
-            storeAppendPrintf(sentry, "%s\t%d\t%d\t%d\n", pool->objectType() + 7, obj_size, pool->getMeter().inuse.level, obj_size * pool->getMeter().inuse.level);
+            storeAppendPrintf(sentry, "%s\t%d\t%ld\t%ld\n", pool->objectType() + 7, obj_size, (long int)pool->getMeter().inuse.level, (long int)obj_size * pool->getMeter().inuse.level);
         }
     }
